==> Programs/40_make_map.Rmd <==
---
title: "40_make_map"
author: "Addison Larson"
date: "Started 20221122 // Last knitted `r format(Sys.time(), '%Y%m%d')`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float: true
    theme: yeti
knit: (function(inputFile, encoding) {
      rmarkdown::render(inputFile,
                        encoding = encoding,
                        output_dir = "C:/Users/ALarson/My_Projects/chi_town/Programs",
                        output_file = "40_make_map.html") })
---

**Purpose:** Show results in a map.

# Setup

```{r setup}

knitr::opts_chunk$set(echo = TRUE)

library(dplyr)
library(sf)
library(here)
library(leaflet)
library(shiny)
library(htmlwidgets)
library(RColorBrewer)
library(htmltools)
library(htmlwidgets)

root     <- here()
analysis <- file.path(root, "Data/Analysis") 

data <- st_read(file.path(analysis, "analysisfile_score.gpkg")) %>%
  st_transform("+proj=longlat +datum=WGS84") %>%
  mutate(
    tract_pct_singlefamily = tract_pct_singlefamily * 100,
    tract_pct_multifamily = tract_pct_multifamily * 100,
  )

outfile <- file.path(root, "Results/map.html")

add_leaflet_layer_univariate <- function(
    map,                  # Name of existing leaflet object
    data_spatial,         # Spatial df with variable and geometries to map
    idvar,                # ID variable of spatial df
    var,                  # Variable to map
    cats_max,             # Max value of each category in choropleth. Supply as list.
    # Right now category is NOT inclusive of this value (e.g. for cat1, if cat_max is 5, then highest value in cat is 4.999)
    # Except for max of highest category, of course
    labs,                 # Value label of each category in choropleth. Supply as list.
    colors,               # Hex color code of each category in choropleth. Supply as list.
    na_color = "#A9A9A9", # Color when data is missing
    layer_name            # Descriptive name of variable to map; is stored as map layer name
) {
  
  # Check that cats_max, labs, and colors have identical lengths
  if ((length(cats_max) != length(labs)) | (length(cats_max) != length(colors))) {
    lengths <- sort(unique(c(length(cats_max), length(labs), length(colors))))
    stop("Arguments cats_max, labs, and colors must be identical lengths. Right now they are lengths: ", paste(lengths, collapse = ", "))
  }
  
  if (length(cats_max) < 2) {
    stop("Can't make a choropleth map with fewer than 2 classes of data. Arguments cats_max, labs, and colors must be length 2 or longer.")
  }
  
  cats_max <- unlist(cats_max)
  labs     <- unlist(labs)
  colors   <- unlist(colors)
  
  # Compute bounding box
  box <- as.numeric(st_bbox(data_spatial))
  
  # Create variable on data to display custom categories
  n_cats <- length(cats_max)
  data_flat <- st_set_geometry(data_spatial, NULL)
  data_flat$var <- data_flat[,var]
  
  # Fill in lowest category
  data_flat$cat <- ifelse(data_flat[,var] < cats_max[1], labs[1], NA_character_)
  
  for (i in 2:length(cats_max)) {
    if (i < length(cats_max)) {
      # Fill in all but top category
      data_flat$cat <- ifelse(data_flat[,var] >= cats_max[i - 1] & data_flat[,var] < cats_max[i], labs[i], data_flat$cat)
    } else {
      # Fill in top category (note no ceiling value)
      data_flat$cat <- ifelse(data_flat[,var] >= cats_max[i - 1], labs[i], data_flat$cat)
    }
  }
  
  # Fill in NAs
  data_flat$cat <- ifelse(is.na(data_flat[,var]), "Data not available", data_flat$cat)
  
  # Add colors
  colors_fn <- colorFactor(levels = c(labs, "Data not available"),
                           palette = c(colors, na_color), na.color = na_color)
  
  data_flat$cat_color <- colors_fn(data_flat$cat)
  
  map_data <- left_join(data_spatial %>% select(!!idvar, geom), data_flat)
  
  map <- map %>%
    fitBounds(box[1], box[2], box[3], box[4]) %>%
    addPolygons(data = map_data,
                group = layer_name,
                fillColor = map_data$cat_color,
                fillOpacity = 0.8,
                smoothFactor = 1,
                stroke = TRUE,
                weight = 0.5,
                color = "#000000",
                label = paste0(
                  "<b> GEOID ", data$geoid, "</b><br>",
                  "Neighborhood: ", data$neighborhood, "<br>",
                  "Overall score: ", sprintf("%.1f", data$overall_score), "<br>",
                  "Transit time: ", sprintf("%.0f", data$transit_duration), "<br>",
                  "Walk distance: ", sprintf("%.0f", data$walk_distance), "<br>",
                  "Median monthly rent: ", sprintf("%.0f", data$median_rent_est), "<br>",
                  "Distance to nearest library: ", sprintf("%.1f", data$library_mindist_topcode), "<br>",
                  "Count of nearby grocery stores: ", sprintf("%.0f", data$n_grocery_0.5mi), "<br>",
                  "Distance to nearest park: ", sprintf("%.1f", data$parks_mindist_topcode), "<br>",
                  "Pct. of tract land, single-family residential: ", sprintf("%.1f", data$tract_pct_singlefamily), "<br>",
                  "Pct. of tract land, multi-family residential: ", sprintf("%.1f", data$tract_pct_multifamily), "<br>"
                ) %>% lapply(htmltools::HTML),
                labelOptions = labelOptions(
                  style = list("font-weight" = "normal", padding = "3 px 8 px"),
                  textsize = "15px",
                  direction = "auto"),
                highlightOptions = highlightOptions(color = "black",
                                                    weight =  2,
                                                    bringToFront = TRUE)) %>%
    addLegend("bottomright",
              group = layer_name,
              colors = c(colors, na_color),
              labels = c(labs, "Data not available"),
              title = layer_name,
              opacity = 1)
  
  return(map)
  
}

```

# Make map

```{r make_map}

map <- leaflet() %>%
  addProviderTiles(providers$Stamen.TonerLite)

summary(data$overall_score)

map <-
  add_leaflet_layer_univariate(
    map = map,
    data_spatial = data,
    idvar = "geoid",
    var = "overall_score",
    cats_max = list(0.3, 0.4, 0.5, 0.6, Inf),
    labs = c("0-0.29", "0.3-0.39", "0.4-0.49", "0.5-0.59", "0.6 or more"),
    colors = brewer.pal(5, "PuBu"),
    layer_name = "Overall score"
  )
map <-
  add_leaflet_layer_univariate(
    map = map,
    data_spatial = data,
    idvar = "geoid",
    var = "transit_duration",
    cats_max = list(15, 30, 45, 60, Inf),
    labs = c("0-14", "15-29", "30-44", "45-59", "60 or more"),
    colors = rev(brewer.pal(5, "PuBu")),
    layer_name = "Transit duration (minutes)"
  )
map <-
  add_leaflet_layer_univariate(
    map = map,
    data_spatial = data,
    idvar = "geoid",
    var = "walk_distance",
    cats_max = list(1, 2, 5, 10, Inf),
    labs = c("0-0.9", "1-1.9", "2-4.9", "5-9.9", "10 or more"),
    colors = rev(brewer.pal(5, "PuBu")),
    layer_name = "Walk distance (miles)"
  )
map <-
  add_leaflet_layer_univariate(
    map = map,
    data_spatial = data,
    idvar = "geoid",
    var = "median_rent_est",
    cats_max = list(750, 1250, 1500, 2000, Inf),
    labs = c("0-749", "750-1,249", "1,250-1,499", "1,500-1,999", "2,000 or more"),
    colors = rev(brewer.pal(5, "RdPu")),
    layer_name = "Median monthly rent ($)"
  )
map <-
  add_leaflet_layer_univariate(
    map = map,
    data_spatial = data,
    idvar = "geoid",
    var = "library_mindist_topcode",
    cats_max = list(0.25, 0.5, 1, Inf),
    labs = c("0-0.24", "0.25-0.49", "0.5-0.9", "1 or more"),
    colors = rev(brewer.pal(4, "GnBu")),
    layer_name = "Distance to nearest library (miles)"
  )
map <-
  add_leaflet_layer_univariate(
    map = map,
    data_spatial = data,
    idvar = "geoid",
    var = "n_grocery_0.5mi",
    cats_max = list(0.001, 1.001, 2.001, 3.001, Inf),
    labs = c("0", "1", "2", "3", "4 or more"),
    colors = brewer.pal(5, "GnBu"),
    layer_name = "Count of grocery stores within 0.5 miles"
  )
map <-
  add_leaflet_layer_univariate(
    map = map,
    data_spatial = data,
    idvar = "geoid",
    var = "parks_mindist_topcode",
    cats_max = list(0.1, 0.25, 0.5, 1, Inf),
    labs = c("0", "0.1-0.24", "0.25-0.49", "0.5-0.99", "1 or more"),
    colors = rev(brewer.pal(5, "Greens")),
    layer_name = "Distance to nearest park (miles)"
  )
map <-
  add_leaflet_layer_univariate(
    map = map,
    data_spatial = data,
    idvar = "geoid",
    var = "tract_pct_singlefamily",
    cats_max = list(10, 20, 30, 40, Inf),
    labs = c("0-9.9%", "10-19.9%", "20-29.9%", "30-39.9%", "40% or more"),
    colors = brewer.pal(5, "YlGnBu"),
    layer_name = "Pct. of tract land area, single-family residential"
  )
map <-
  add_leaflet_layer_univariate(
    map = map,
    data_spatial = data,
    idvar = "geoid",
    var = "tract_pct_multifamily",
    cats_max = list(10, 20, 30, 40, Inf),
    labs = c("0-9.9%", "10-19.9%", "20-29.9%", "30-39.9%", "40% or more"),
    colors = brewer.pal(5, "YlGnBu"),
    layer_name = "Pct. of tract land area, multi-family residential"
  )

map <- addLayersControl(
  map,
  overlayGroups = c(
    "Overall score",
    "Transit duration (minutes)",
    "Walk distance (miles)",
    "Median monthly rent ($)",
    "Distance to nearest library (miles)",
    "Count of grocery stores within 0.5 miles",
    "Distance to nearest park (miles)",
    "Pct. of tract land area, single-family residential",
    "Pct. of tract land area, multi-family residential"
  ),
  options = layersControlOptions(collapsed = FALSE, autoZIndex = TRUE)
) %>%
  hideGroup(c(
    "Transit duration (minutes)",
    "Walk distance (miles)",
    "Median monthly rent ($)",
    "Distance to nearest library (miles)",
    "Count of grocery stores within 0.5 miles",
    "Distance to nearest park (miles)",
    "Pct. of tract land area, single-family residential",
    "Pct. of tract land area, multi-family residential"
  ))

map

```

# Export map

```{r export}

saveWidget(map, outfile)

```
